move examples to a standalone file
